feat(desktop): add Cline as a tier-2 preset harness
Cline ships a standalone CLI (npm i -g cline) with native ACP support
via `cline --acp`, matching the shape of the other flag-preset
harnesses (opencode, kimi, cursor). No adapter needed.

Split the tier-2 PresetHarness block (struct, PRESET_HARNESSES, and
its two accessor fns) out of discovery.rs into a new sibling module,
discovery/preset_harnesses.rs. discovery.rs had already reached the
desktop file-size ratchet (1860 lines); any further PresetHarness
addition — this one included — failed the ratchet check. The split
gives both files headroom again for future presets.
